LPS-121803 portal-search-elasticsearch7-impl: Encapsulate divergences in bulkable document translator subsystem
+ (improves source code parity across versions)

https://issues.liferay.com/browse/LPS-121803
