+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/search/engine/adapter/document/BulkDocumentRequestExecutorImpl.java
@@ -22,7 +22,6 @@ import com.liferay.portal.search.engine.adapter.document.BulkDocumentItemRespons
 import com.liferay.portal.search.engine.adapter.document.BulkDocumentRequest;
 import com.liferay.portal.search.engine.adapter.document.BulkDocumentResponse;
 import com.liferay.portal.search.engine.adapter.document.BulkableDocumentRequest;
-import com.liferay.portal.search.engine.adapter.document.BulkableDocumentRequestTranslator;
 import com.liferay.portal.search.engine.adapter.document.DeleteDocumentRequest;
 import com.liferay.portal.search.engine.adapter.document.IndexDocumentRequest;
 import com.liferay.portal.search.engine.adapter.document.UpdateDocumentRequest;
@@ -118,22 +117,22 @@ public class BulkDocumentRequestExecutorImpl
 				request -> {
 					if (request instanceof DeleteDocumentRequest) {
 						DeleteRequest deleteRequest =
-							_bulkableDocumentRequestTranslator.translate(
-								(DeleteDocumentRequest)request);
+							_elasticsearchBulkableDocumentRequestTranslator.
+								translate((DeleteDocumentRequest)request);
 
 						bulkRequest.add(deleteRequest);
 					}
 					else if (request instanceof IndexDocumentRequest) {
 						IndexRequest indexRequest =
-							_bulkableDocumentRequestTranslator.translate(
-								(IndexDocumentRequest)request);
+							_elasticsearchBulkableDocumentRequestTranslator.
+								translate((IndexDocumentRequest)request);
 
 						bulkRequest.add(indexRequest);
 					}
 					else if (request instanceof UpdateDocumentRequest) {
 						UpdateRequest updateRequest =
-							_bulkableDocumentRequestTranslator.translate(
-								(UpdateDocumentRequest)request);
+							_elasticsearchBulkableDocumentRequestTranslator.
+								translate((UpdateDocumentRequest)request);
 
 						bulkRequest.add(updateRequest);
 					}
@@ -165,10 +164,12 @@ public class BulkDocumentRequestExecutorImpl
 	}
 
 	@Reference(target = "(search.engine.impl=Elasticsearch)", unbind = "-")
-	protected void setBulkableDocumentRequestTranslator(
-		BulkableDocumentRequestTranslator bulkableDocumentRequestTranslator) {
+	protected void setElasticsearchBulkableDocumentRequestTranslator(
+		ElasticsearchBulkableDocumentRequestTranslator
+			elasticsearchBulkableDocumentRequestTranslator) {
 
-		_bulkableDocumentRequestTranslator = bulkableDocumentRequestTranslator;
+		_elasticsearchBulkableDocumentRequestTranslator =
+			elasticsearchBulkableDocumentRequestTranslator;
 	}
 
 	@Reference(unbind = "-")
@@ -181,8 +182,8 @@ public class BulkDocumentRequestExecutorImpl
 	private static final Log _log = LogFactoryUtil.getLog(
 		BulkDocumentRequestExecutorImpl.class);
 
-	private BulkableDocumentRequestTranslator
-		_bulkableDocumentRequestTranslator;
+	private ElasticsearchBulkableDocumentRequestTranslator
+		_elasticsearchBulkableDocumentRequestTranslator;
 	private ElasticsearchClientResolver _elasticsearchClientResolver;
 
 }
