+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/test/java/com/liferay/portal/search/elasticsearch7/internal/search/engine/adapter/document/DocumentRequestExecutorFixture.java
@@ -17,7 +17,6 @@ package com.liferay.portal.search.elasticsearch7.internal.search.engine.adapter.
 import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchClientResolver;
 import com.liferay.portal.search.elasticsearch7.internal.document.ElasticsearchDocumentFactory;
 import com.liferay.portal.search.elasticsearch7.internal.query.ElasticsearchQueryTranslatorFixture;
-import com.liferay.portal.search.engine.adapter.document.BulkableDocumentRequestTranslator;
 import com.liferay.portal.search.engine.adapter.document.DocumentRequestExecutor;
 import com.liferay.portal.search.internal.document.DocumentBuilderFactoryImpl;
 import com.liferay.portal.search.internal.geolocation.GeoBuildersImpl;
@@ -38,11 +37,11 @@ public class DocumentRequestExecutorFixture {
 			_elasticsearchClientResolver, _elasticsearchDocumentFactory);
 	}
 
-	protected static BulkableDocumentRequestTranslator
+	protected static ElasticsearchBulkableDocumentRequestTranslator
 		createBulkableDocumentRequestTranslator(
 			ElasticsearchDocumentFactory elasticsearchDocumentFactory) {
 
-		return new ElasticsearchBulkableDocumentRequestTranslator() {
+		return new ElasticsearchBulkableDocumentRequestTranslatorImpl() {
 			{
 				setElasticsearchDocumentFactory(elasticsearchDocumentFactory);
 			}
@@ -52,12 +51,12 @@ public class DocumentRequestExecutorFixture {
 	protected static BulkDocumentRequestExecutor
 		createBulkDocumentRequestExecutor(
 			ElasticsearchClientResolver elasticsearchClientResolver,
-			BulkableDocumentRequestTranslator
+			ElasticsearchBulkableDocumentRequestTranslator
 				bulkableDocumentRequestTranslator) {
 
 		return new BulkDocumentRequestExecutorImpl() {
 			{
-				setBulkableDocumentRequestTranslator(
+				setElasticsearchBulkableDocumentRequestTranslator(
 					bulkableDocumentRequestTranslator);
 				setElasticsearchClientResolver(elasticsearchClientResolver);
 			}
@@ -96,7 +95,7 @@ public class DocumentRequestExecutorFixture {
 	protected static DeleteDocumentRequestExecutor
 		createDeleteDocumentRequestExecutor(
 			ElasticsearchClientResolver elasticsearchClientResolver,
-			BulkableDocumentRequestTranslator
+			ElasticsearchBulkableDocumentRequestTranslator
 				bulkableDocumentRequestTranslator) {
 
 		return new DeleteDocumentRequestExecutorImpl() {
@@ -112,9 +111,10 @@ public class DocumentRequestExecutorFixture {
 		ElasticsearchClientResolver elasticsearchClientResolver,
 		ElasticsearchDocumentFactory elasticsearchDocumentFactory) {
 
-		BulkableDocumentRequestTranslator bulkableDocumentRequestTranslator =
-			createBulkableDocumentRequestTranslator(
-				elasticsearchDocumentFactory);
+		ElasticsearchBulkableDocumentRequestTranslator
+			bulkableDocumentRequestTranslator =
+				createBulkableDocumentRequestTranslator(
+					elasticsearchDocumentFactory);
 
 		return new ElasticsearchDocumentRequestExecutor() {
 			{
@@ -151,7 +151,7 @@ public class DocumentRequestExecutorFixture {
 	protected static GetDocumentRequestExecutor
 		createGetDocumentRequestExecutor(
 			ElasticsearchClientResolver elasticsearchClientResolver,
-			BulkableDocumentRequestTranslator
+			ElasticsearchBulkableDocumentRequestTranslator
 				bulkableDocumentRequestTranslator) {
 
 		return new GetDocumentRequestExecutorImpl() {
@@ -168,7 +168,7 @@ public class DocumentRequestExecutorFixture {
 	protected static IndexDocumentRequestExecutor
 		createIndexDocumentRequestExecutor(
 			ElasticsearchClientResolver elasticsearchClientResolver,
-			BulkableDocumentRequestTranslator
+			ElasticsearchBulkableDocumentRequestTranslator
 				bulkableDocumentRequestTranslator) {
 
 		return new IndexDocumentRequestExecutorImpl() {
@@ -214,7 +214,7 @@ public class DocumentRequestExecutorFixture {
 	protected static UpdateDocumentRequestExecutor
 		createUpdateDocumentRequestExecutor(
 			ElasticsearchClientResolver elasticsearchClientResolver,
-			BulkableDocumentRequestTranslator
+			ElasticsearchBulkableDocumentRequestTranslator
 				bulkableDocumentRequestTranslator) {
 
 		return new UpdateDocumentRequestExecutorImpl() {
